+++ actix-settings/CHANGES.md
@@ -2,6 +2,15 @@
 
 ## Unreleased
 
+## 0.7.1
+
+- Fix doc examples.
+
+## 0.7.0
+
+- The `ApplySettings` trait now includes a type parameter, allowing multiple types to be implemented per configuration target.
+- Implement `ApplySettings` for `ActixSettings`.
+- `BasicSettings::from_default_template()` is now infallible.
 - Rename `AtError => Error`.
 - Remove `AtResult` type alias.
 - Update `toml` dependency to `0.8`.